# specs/001-purple-glow-social/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Purple Glow Social API
  version: 1.0.0
  description: |
    RESTful API for Purple Glow Social - AI-powered social media management platform.

    **Authentication**: All endpoints (except /auth/*) require Bearer token in Authorization header.

    **Rate Limiting**: 
    - Authenticated: 100 req/min per user
    - Unauthenticated: 10 req/min per IP

    **Error Responses**: All endpoints return RFC 7807 Problem Details for HTTP APIs format.
  contact:
    name: Purple Glow Social Support
    email: support@purpleglowsocial.com
  license:
    name: Proprietary
servers:
  - url: https://api.purpleglowsocial.com/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase Auth
  schemas:
    Error:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          example: 'https://api.purpleglowsocial.com/errors/validation-error'
        title:
          type: string
          example: 'Validation Error'
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: 'The request body contains invalid data.'
        instance:
          type: string
          format: uri
          example: '/api/v1/posts'
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    BusinessProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_user_id:
          type: string
          format: uuid
        business_name:
          type: string
        industry:
          type: string
        target_audience:
          type: string
        brand_logo_url:
          type: string
          format: uri
          nullable: true
        primary_color:
          type: string
          pattern: '^#[0-9A-F]{6}$'
          nullable: true
        secondary_color:
          type: string
          pattern: '^#[0-9A-F]{6}$'
          nullable: true
        content_tone:
          type: string
          enum: [professional, casual, friendly, formal, humorous]
        content_topics:
          type: array
          items:
            type: string
        preferred_language:
          type: string
          enum: [en, af, zu, xh, nso, st, ss, ts, tn, ve, nr]
        posting_frequency:
          type: string
          enum: [daily, 3x_week, weekly, custom]
        automation_enabled:
          type: boolean
        automation_eligible_at:
          type: string
          format: date-time
          nullable: true
        approved_posts_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        sender:
          type: string
          enum: [user, system]
        message_text:
          type: string
        interpreted_command:
          type: string
          nullable: true
        command_parameters:
          type: object
          nullable: true
        resulting_action:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
  responses:
    UnauthorizedError:
      description: Authentication token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: 'https://api.purpleglowsocial.com/errors/unauthorized'
            title: 'Unauthorized'
            status: 401
            detail: 'Invalid or expired authentication token.'
...
    NotFoundError:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: 'https://api.purpleglowsocial.com/errors/not-found'
            title: 'Not Found'
            status: 404
            detail: 'The requested resource does not exist.'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: 'https://api.purpleglowsocial.com/errors/validation-error'
            title: 'Validation Error'
            status: 400
            detail: 'The request body contains invalid data.'
            errors:
              - field: 'business_name'
                message: 'Business name must be at least 2 characters.'
...
security:
  - BearerAuth: []
paths:
...
  /business-profiles/me:
    get:
      tags:
        - Business Profiles
      summary: Get current user's business profile
      responses:
        '200':
          description: Business profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessProfile'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Business Profiles
      summary: Update business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                business_name:
                  type: string
                industry:
                  type: string
                target_audience:
                  type: string
                brand_logo_url:
                  type: string
                  format: uri
                primary_color:
                  type: string
                secondary_color:
                  type: string
                content_tone:
                  type: string
                content_topics:
                  type: array
                  items:
                    type: string
                preferred_language:
                  type: string
                posting_frequency:
                  type: string
      responses:
        '200':
          description: Business profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessProfile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /chat/messages:
    get:
      tags:
        - Chat
      summary: Get chat message history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Messages retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessage'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Chat
      summary: Send chat message
      description: Send user message and get system response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message_text
              properties:
                message_text:
                  type: string
      responses:
        '201':
          description: Message sent and response generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_message:
                    $ref: '#/components/schemas/ChatMessage'
                  system_response:
                    $ref: '#/components/schemas/ChatMessage'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ========================================
  # Admin Endpoints
  # ========================================
...